+++ flake.nix
@@ -142,6 +142,15 @@
       };
     }
     {
+      host = "sigma";
+      user = "enzime";
+      system = "x86_64-linux";
+      nixos = true;
+      modules = builtins.attrNames {
+        inherit (modules) i3 laptop;
+      };
+    }
+    {
       host = "tau";
       user = "enzime";
       system = "x86_64-linux";