+++ keys.nix
@@ -1,12 +1,14 @@
 {
   users = {
     enzime_phi = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAILzD/3cOhlqe8NVEruSUnPSnG1GbmX8SgTbVGLFHMa7g";
+    enzime_sigma = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIHwM0sVLdolDXb7LyKMT6Jfym0dP1NSgJClHcIOga+fa";
     enzime_tau = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAICqDFjlA1zdrYy9lyv9TQrXJF89fgxcWRZSiU/xDoog/";
     nathan = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIHZcON5uabdGPUp5Sf161wShGEwhNklD8w50f6EnkPgo";
   };
 
   hosts = {
     phi = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMxOi/S1TLBg8/ZRX5XfCTlM8A+I0q0pQksrxtfjdYFP";
+    sigma = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIN3pv/iYVewBlpTImGMvChu7da6jsi99/Y0vpsmAoeds";
     tau = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIP4Yk5eL/z/I9KhXm1ZsOBQiIBgdGZe1xUTz8whBWe4u";
     apollo = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIDVVHXZq7684xQZGSLGmHiSCbl+WjfkZGgzk1Kz28nXY";
   };