+++ modules/wireless.nix
@@ -1,0 +1,5 @@
+{
+  nixosModule = { config, ... }: {
+    networking.networkmanager.enable = true;
+  };
+}